use JUnit for `ConcreteCollectionCheck` tests
