--- autograder-core/src/test/resources/de/firemage/autograder/core/check_tests/ConcreteCollectionCheck/config.txt
@@ -1,23 +1,0 @@
-oop.ConcreteCollectionCheck
-Program against collection interfaces, not concrete types
-Test.java:10
-Test.java:13
-Test.java:16
-Test.java:19
-Test.java:20
-Test.java:33
-Test.java:43
-Test.java:53
-Test.java:63
-Test.java:64
-Test.java:66
-Test.java:80
-Test.java:81
-Test.java:82
-Test.java:106
-Test.java:131
-Test.java:136
-Test.java:140
-Test.java:154
-Test.java:162
-Test.java:174
